refactor: payment 도메인 리팩토링 (#35)
* refactor: transaction 엔티티의 name 필드를 depositorName으로 변경

* refactor: depositorName 필드 인덱스 설정

* refactor: transaction과 payment 사이의 연관 관계 제거

* refactor: transaction 엔티티의 금액 관련 필드의 타입을 BigDecimal로 변경

* feat: 결제 취소 상태 추가

* refactor: 현재 입금 금액을 payment 객체가 아닌 transaction 테이블에서 합산하여 가져오도록 수정

* refactor: 트랜잭션 합계 메서드가 null을 반환할 수 없도록 반환 타입을 Optional로 변경

* feat: 사용자가 결제 정보 생성을 중복으로 요청할 때 상태에 맞게 엔티티 상태를 업데이트

* fix: 엔티티에 맞게 반환 타입을 변경

* fix: 엔티티에 맞게 테스트 코드를 조정

* test: 현재 입금된 금액을 가져오는 Helper 메서드 추가

* refactor: 결제 상태를 거래내역에 맞춰 업데이트

+++ src/main/java/aegis/server/domain/payment/service/parser/IbkTransactionParser.java
@@ -4,6 +4,7 @@ import aegis.server.domain.payment.domain.Transaction;
 import aegis.server.domain.payment.domain.TransactionType;
 import lombok.RequiredArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.Clock;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
@@ -40,7 +41,7 @@ public class IbkTransactionParser implements TransactionParser {
             default -> throw new IllegalArgumentException("알 수 없는 거래유형입니다");
         };
 
-        Long amount = Long.parseLong(matcher.group(2).replace(",", ""));
+        BigDecimal amount = BigDecimal.valueOf(Long.parseLong(matcher.group(2).replace(",", "")));
 
         String name = matcher.group(3);
 
@@ -64,7 +65,7 @@ public class IbkTransactionParser implements TransactionParser {
                 ? parsedTime.minusYears(1)
                 : parsedTime;
 
-        Long balance = Long.parseLong(matcher.group(2).replace(",", ""));
+        BigDecimal balance = BigDecimal.valueOf(Long.parseLong(matcher.group(2).replace(",", "")));
 
         return Transaction.of(transactionTime, name, transactionType, amount, balance);
     }
